fix: 지하철 api 호출 시 라우트 활용하도록 변경
